+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Mui Alert for unauthorized users on creator cards
 - Mui lab package to use experimental components currently in v5 of core package
 
 ### Changed